Tue, Jun 18, 2024, 9:55 PM +05:30

+++ Lecture-1-C/Practice Set/PS2.c
@@ -5,7 +5,7 @@ int main(void) {
 
     printf("Fahrenheit to Celcius.\n");
 
-    int F = get_int("F:");
+    float F = get_int("F:");
     float out = (F - 32) * 5 / 9;
 
     printf("%f Celcius\n", out);
